Front end pages for Login, Dashboard and Settings
