+++ HtmlFiles/dashboard.html
@@ -1,0 +1,106 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>TheWidgetWebsite</title>
+    <!-- CSS link -->
+    <link rel="stylesheet" href="../CSS/dashboard.css" />
+
+    <!-- JS link -->
+    <script defer src="../JavaScript/widget.js"></script>
+  </head>
+  <body>
+    <header class="p-3 mb-3 border-bottom">
+      <div class="container">
+        <div
+          class="d-flex flex-wrap align-items-center justify-content-center justify-content-lg-start"
+        >
+          <a
+            href="/"
+            class="d-flex align-items-center mb-2 mb-lg-0 text-dark text-decoration-none"
+          >
+            <svg
+              class="bi me-2"
+              width="40"
+              height="32"
+              role="img"
+              aria-label="Bootstrap"
+            >
+              <use xlink:href="#bootstrap"></use>
+            </svg>
+          </a>
+
+          <ul
+            class="nav col-12 col-lg-auto me-lg-auto mb-2 justify-content-center mb-md-0"
+          >
+            <li>
+              <a href="#" class="nav-link px-2 link-secondary">Overview</a>
+            </li>
+            <li><a href="#" class="nav-link px-2 link-dark">LogIn</a></li>
+            <li><a href="#" class="nav-link px-2 link-dark">Dashboard</a></li>
+            <li><a href="#" class="nav-link px-2 link-dark">Settings</a></li>
+          </ul>
+
+          <form class="col-12 col-lg-auto mb-3 mb-lg-0 me-lg-3" role="search">
+            <input
+              type="search"
+              class="form-control"
+              placeholder="Search..."
+              aria-label="Search"
+            />
+          </form>
+
+          <div class="dropdown text-end">
+            <a
+              href="#"
+              class="d-block link-dark text-decoration-none dropdown-toggle"
+              id="dropdownUser1"
+              data-bs-toggle="dropdown"
+              aria-expanded="false"
+            >
+              <img
+                src="../imagesXicons/pic1.png"
+                alt="mdo"
+                width="40"
+                height="40"
+                class="rounded-circle"
+              />
+            </a>
+            <ul
+              class="dropdown-menu text-small"
+              aria-labelledby="dropdownUser1"
+            >
+              <li><a class="dropdown-item" href="#">New project...</a></li>
+              <li><a class="dropdown-item" href="#">Settings</a></li>
+              <li><a class="dropdown-item" href="#">Profile</a></li>
+              <li><hr class="dropdown-divider" /></li>
+              <li><a class="dropdown-item" href="#">Sign out</a></li>
+            </ul>
+          </div>
+        </div>
+      </div>
+    </header>
+
+    <div
+      id="profile_pic"
+      class="container mt-4 mb-4 p-3 d-flex justify-content-center"
+    >
+      <div class="card p-4">
+        <div
+          class="image d-flex flex-column justify-content-center align-items-center"
+        >
+          <button class="btn btn-secondary">
+            <img src="../imagesXicons/pic1.png" height="100" width="100" />
+          </button>
+          <span class="name mt-3">Test McTesty</span>
+          <span class="idd">@test</span>
+          <div class="d-flex mt-2">
+            <button class="btn1 btn-dark">Edit Profile</button>
+          </div>
+        </div>
+      </div>
+    </div>
+  </body>
+</html>
