+++ HtmlFiles/login.html
@@ -1,0 +1,86 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>TheWidgetWebsite</title>
+    <!-- CSS link -->
+    <link rel="stylesheet" href="../CSS/login.css" />
+
+    <!-- JS link -->
+    <script defer src="../JavaScript/widget.js"></script>
+  </head>
+  <body>
+    <header>
+      <h1 id="siteTitle">TheWidgetWebsite</h1>
+    </header>
+    <div id="widget_container" class="float_container">
+      <div id="icon_container" class="float_child">
+        <p>This container contains an image and icon files</p>
+      </div>
+      <div id="form_container" class="float_child">
+        <form id="create_form" action="result.html" method="GET">
+          <div class="form-group">
+            <label>Please enter your name</label>
+            <input
+              type="text"
+              class="form-control"
+              id="username"
+              name="user_name"
+              required
+            />
+          </div>
+          <div class="form-group">
+            <label>Please enter your Email</label>
+            <input
+              type="password"
+              class="form-control"
+              id="password"
+              name="user_password"
+              required
+            />
+          </div>
+
+          <div class="form-row">
+            <div class="form-group col-md-6">
+              <label>Email</label>
+              <input
+                type="email"
+                class="form-control"
+                id="email"
+                name="user_email"
+              />
+            </div>
+          </div>
+
+          <div class="custom-control custom-checkbox my-1 mr-sm-2">
+            <input
+              type="checkbox"
+              class="custom-control-input"
+              id="customControlInline"
+              name="check_box"
+            />
+            <label class="custom-control-label" for="customControlInline"
+              >Remember password</label
+            >
+          </div>
+          <div class="row">
+            <div class="col-md-12">
+              <button
+                type="submit"
+                class="btn btn-primary btn-md btn-block waves-effect text-center m-b-20"
+                name="btnn"
+              >
+                <i class="fa fa-lock"></i>Log In
+              </button>
+            </div>
+          </div>  
+        </div>
+      </form>
+    </div>
+    <footer>
+      <p id="footer"><strong>insert footer link notes</strong></p>
+    </footer>
+  </body>
+</html>
